+++ src/Builder/BuilderMetaBox.php
@@ -78,7 +78,7 @@ class BuilderMetaBox {
 		<input type="hidden" id="field-data" value="<?php echo esc_attr( wp_json_encode( $fields ) ); ?>">
 
 		<div id="puppyfw-builder">
-			<input type="hidden" name="puppyfw_fields" :value="JSON.stringify(fields)">
+			<input type="hidden" id="puppyfw-field-save-data" name="puppyfw_fields" :value="JSON.stringify(fields)">
 
 			<fields-builder
 				:fields="fields"